Refactor the binning scheme generation code to be object-oriented. Users can define a derived class from BinSchemeBase and enable it to be automatically constructed by BinSchemeFactory.
