+++ configs/tutorial_slice_config.txt
@@ -1,39 +1,12 @@
 3
-"reco p_{p}" "GeV/c" "reco $p_{\mu}$" "GeV$/c$"
-"reco cos#theta_{p}" "" "reco $\cos\theta_{\mu}$" ""
-"reco bin number" "" "reco bin number" ""
+"muon cos#theta" "" "\cos\theta_{\mu}" ""
+"p_{#mu}" " (GeV)" "p_{\mu}" " (GeV)"
+"bin number" "bin number" "" ""
 3
 "events"
-1 1 10 -1 -0.5 0 0.27 0.45 0.62 0.76 0.86 0.94 1
+1 0 31 -1 -0.85 -0.775 -0.7 -0.625 -0.55 -0.475 -0.4 -0.325 -0.25 -0.175 -0.1 -0.025 0.05 0.125 0.2 0.275 0.35 0.425 0.5 0.575 0.65 0.725 0.8 0.85 0.875 0.9 0.925 0.95 0.975 1
 0
-9
-0 1 1
-1 1 2
-2 1 3
-3 1 4
-4 1 5
-5 1 6
-6 1 7
-7 1 8
-8 1 9
-"events"
-1 0 11 0.25 0.41 0.495 0.56 0.62 0.68 0.74 0.8 0.87 0.93 1.2
-0
-10
-9 1 1
-10 1 2
-11 1 3
-12 1 4
-13 1 5
-14 1 6
-15 1 7
-16 1 8
-17 1 9
-18 1 10
-"events"
-1 2 20 0 1 2 3 4 5 6 7 8 9 10 11 12 13 14 15 16 17 18 19
-0
-19
+30
 0 1 1
 1 1 2
 2 1 3
@@ -53,3 +26,104 @@
 16 1 17
 17 1 18
 18 1 19
+19 1 20
+20 1 21
+21 1 22
+22 1 23
+23 1 24
+24 1 25
+25 1 26
+26 1 27
+27 1 28
+28 1 29
+29 1 30
+"events"
+1 1 27 0.1 0.175 0.2 0.225 0.25 0.275 0.3 0.325 0.35 0.375 0.4 0.425 0.45 0.475 0.5 0.55 0.6 0.65 0.7 0.75 0.8 0.85 0.9 0.95 1 1.1 1.2
+0
+26
+30 1 1
+31 1 2
+32 1 3
+33 1 4
+34 1 5
+35 1 6
+36 1 7
+37 1 8
+38 1 9
+39 1 10
+40 1 11
+41 1 12
+42 1 13
+43 1 14
+44 1 15
+45 1 16
+46 1 17
+47 1 18
+48 1 19
+49 1 20
+50 1 21
+51 1 22
+52 1 23
+53 1 24
+54 1 25
+55 1 26
+"events"
+1 2 57 0 1 2 3 4 5 6 7 8 9 10 11 12 13 14 15 16 17 18 19 20 21 22 23 24 25 26 27 28 29 30 31 32 33 34 35 36 37 38 39 40 41 42 43 44 45 46 47 48 49 50 51 52 53 54 55 56
+0
+56
+0 1 1
+1 1 2
+2 1 3
+3 1 4
+4 1 5
+5 1 6
+6 1 7
+7 1 8
+8 1 9
+9 1 10
+10 1 11
+11 1 12
+12 1 13
+13 1 14
+14 1 15
+15 1 16
+16 1 17
+17 1 18
+18 1 19
+19 1 20
+20 1 21
+21 1 22
+22 1 23
+23 1 24
+24 1 25
+25 1 26
+26 1 27
+27 1 28
+28 1 29
+29 1 30
+30 1 31
+31 1 32
+32 1 33
+33 1 34
+34 1 35
+35 1 36
+36 1 37
+37 1 38
+38 1 39
+39 1 40
+40 1 41
+41 1 42
+42 1 43
+43 1 44
+44 1 45
+45 1 46
+46 1 47
+47 1 48
+48 1 49
+49 1 50
+50 1 51
+51 1 52
+52 1 53
+53 1 54
+54 1 55
+55 1 56
