+++ include/XSecAnalyzer/Binning/TutorialBinScheme.hh
@@ -1,0 +1,12 @@
+#pragma once
+
+// XSecAnalyzer includes
+#include "XSecAnalyzer/Binning/BinSchemeBase.hh"
+
+class TutorialBinScheme : public BinSchemeBase {
+
+  public:
+
+    TutorialBinScheme();
+    virtual void DefineBlocks() override;
+};
