+++ src/utils/MakeConfig.cxx
@@ -36,60 +36,54 @@
 #include "XSecAnalyzer/Unfolder.hh"
 #include "XSecAnalyzer/WeightHandler.hh"
 
-// Define this static member of the Universe class
-size_t Universe::num_categories_;
+// Initialize the static dummy counter owned by the MakeConfig class. This
+// counter is used to ensure that each automatically-generated histogram has a
+// unique name to use with TTree::Draw()
+int MakeConfig::hist_count = 0;
 
-using namespace std;
-
-//  TODO: maybe it is good to make the constructor to be private.
-
-MakeConfig::MakeConfig(){
+MakeConfig::MakeConfig( const std::string& bin_scheme_name ) {
   RUNS = {1};
+  bin_scheme_name_ = bin_scheme_name;
 }
 
 MakeConfig::~MakeConfig(){
-  cout << "free make config\n";
-  for(int i = 0; i < vect_block.size(); i++){
-    delete vect_block.at(i).block_true_;
-    delete vect_block.at(i).block_reco_;
+  std::cout << "free make config\n";
+  for(int i = 0; i < vect_block->size(); i++){
+    delete vect_block->at(i).block_true_;
+    delete vect_block->at(i).block_reco_;
   }
 }
 
-// Dummy counter used to ensure that each histogram generated by this
-// function has a unique name to use with TTree::Draw()
-int MakeConfig::hist_count = 0;
+void MakeConfig::ResPlots() {
 
-void MakeConfig::ResPlots(){
-
-
-  for(int i = 0; i < vect_block.size(); i++){
+  for(int i = 0; i < vect_block->size(); i++){
     ++hist_count;
-    if(vect_block.at(i).block_true_->Is1D()){
-      make_res_plots( vect_block[i].block_reco_->GetXName(),
-          vect_block[i].block_reco_->GetXTitle(),
-          vect_block[i].block_reco_->GetSelection(),
+    if(vect_block->at(i).block_true_->Is1D()){
+      make_res_plots( vect_block->at(i).block_reco_->GetXName(),
+          vect_block->at(i).block_reco_->GetXTitle(),
+          vect_block->at(i).block_reco_->GetSelection(),
           RUNS,
-          vect_block[i].block_reco_->GetVector(),
+          vect_block->at(i).block_reco_->GetVector(),
           false, false,
           DEFAULT_TRUE_BINS,
-          vect_block[i].block_true_->GetXName(),
-          vect_block[i].block_true_->GetSelection(),
+          vect_block->at(i).block_true_->GetXName(),
+          vect_block->at(i).block_true_->GetSelection(),
           DEFAULT_MC_EVENT_WEIGHT );
     }
     else{
 
       std::vector< TrueBin > true_bins;
       std::vector< RecoBin > reco_bins;
-      for(int j = 0; j < vect_block.at(i).block_true_->GetNBinsX(); j++){
-        double xlow = vect_block.at(i).block_true_->GetBinXLow(j);
-        double xhigh = vect_block.at(i).block_true_->GetBinXHigh(j);
-        for(int k = 0; k < vect_block.at(i).block_true_->GetNBinsY(j); k++){
+      for(int j = 0; j < vect_block->at(i).block_true_->GetNBinsX(); j++){
+        double xlow = vect_block->at(i).block_true_->GetBinXLow(j);
+        double xhigh = vect_block->at(i).block_true_->GetBinXHigh(j);
+        for(int k = 0; k < vect_block->at(i).block_true_->GetNBinsY(j); k++){
           true_bins.emplace_back(
-            vect_block.at(i).block_true_->GetBinDef(j, k),
-            TrueBinType(vect_block.at(i).block_true_->GetBinType()), i );
+            vect_block->at(i).block_true_->GetBinDef(j, k),
+            TrueBinType(vect_block->at(i).block_true_->GetBinType()), i );
           reco_bins.emplace_back(
-            vect_block.at(i).block_reco_->GetBinDef(j, k),
-            RecoBinType(vect_block.at(i).block_reco_->GetBinType()), i );
+            vect_block->at(i).block_reco_->GetBinDef(j, k),
+            RecoBinType(vect_block->at(i).block_reco_->GetBinType()), i );
         }
       }
 
@@ -120,60 +114,60 @@ void MakeConfig::Print(){
   std::vector< RecoBin > reco_bins;
 
 
-  for(int i = 0; i < vect_block.size(); i++){
+  for(int i = 0; i < vect_block->size(); i++){
 
-    if(vect_block.at(i).block_true_->Is1D()){
+    if(vect_block->at(i).block_true_->Is1D()){
 
-      sb.slice_vars_.emplace_back( vect_block.at(i).block_true_->GetXTitle(),
-          vect_block.at(i).block_true_->GetXTexTitle(),
-          vect_block.at(i).block_true_->GetXTitleUnit(),
-          vect_block.at(i).block_true_->GetXTexTitleUnit() );
+      sb.slice_vars_.emplace_back( vect_block->at(i).block_true_->GetXTitle(),
+          vect_block->at(i).block_true_->GetXTexTitle(),
+          vect_block->at(i).block_true_->GetXTitleUnit(),
+          vect_block->at(i).block_true_->GetXTexTitleUnit() );
       int var_idx = find_slice_var_index(
-        vect_block.at(i).block_true_->GetXTitle(), sb.slice_vars_ );
+        vect_block->at(i).block_true_->GetXTitle(), sb.slice_vars_ );
 
-      size_t num_bins = vect_block.at(i).block_true_->GetNBinsX();
+      size_t num_bins = vect_block->at(i).block_true_->GetNBinsX();
 
-      auto& slice = add_slice(sb, vect_block.at(i).block_true_->GetVector(), var_idx);
+      auto& slice = add_slice(sb, vect_block->at(i).block_true_->GetVector(), var_idx);
 
-      for(int j = 0; j < vect_block.at(i).block_true_->GetNBinsX(); j++){
+      for(int j = 0; j < vect_block->at(i).block_true_->GetNBinsX(); j++){
         slice.bin_map_[ j + 1 ].insert( true_bins.size() );
-        true_bins.emplace_back( vect_block.at(i).block_true_->GetBinDef(j),
-          TrueBinType(vect_block.at(i).block_true_->GetBinType()), i );
+        true_bins.emplace_back( vect_block->at(i).block_true_->GetBinDef(j),
+          TrueBinType(vect_block->at(i).block_true_->GetBinType()), i );
 
       }
-      for(int j = 0; j < vect_block.at(i).block_reco_->GetNBinsX(); j++){
-        reco_bins.emplace_back( vect_block.at(i).block_reco_->GetBinDef(j),
-          RecoBinType(vect_block.at(i).block_reco_->GetBinType()), i );
+      for(int j = 0; j < vect_block->at(i).block_reco_->GetNBinsX(); j++){
+        reco_bins.emplace_back( vect_block->at(i).block_reco_->GetBinDef(j),
+          RecoBinType(vect_block->at(i).block_reco_->GetBinType()), i );
       }
     }
     else{
-      sb.slice_vars_.emplace_back( vect_block.at(i).block_true_->GetXTitle(),
-          vect_block.at(i).block_true_->GetXTexTitle(),
-          vect_block.at(i).block_true_->GetXTitleUnit(),
-          vect_block.at(i).block_true_->GetXTexTitleUnit() );
+      sb.slice_vars_.emplace_back( vect_block->at(i).block_true_->GetXTitle(),
+          vect_block->at(i).block_true_->GetXTexTitle(),
+          vect_block->at(i).block_true_->GetXTitleUnit(),
+          vect_block->at(i).block_true_->GetXTexTitleUnit() );
       sb.slice_vars_.emplace_back(
-        vect_block.at(i).block_true_->GetYTitle(),
-        vect_block.at(i).block_true_->GetYTexTitle(),
-        vect_block.at(i).block_true_->GetYTitleUnit(),
-        vect_block.at(i).block_true_->GetYTexTitleUnit() );
+        vect_block->at(i).block_true_->GetYTitle(),
+        vect_block->at(i).block_true_->GetYTexTitle(),
+        vect_block->at(i).block_true_->GetYTitleUnit(),
+        vect_block->at(i).block_true_->GetYTexTitleUnit() );
 
 
       int xvar_idx = find_slice_var_index(
-        vect_block.at(i).block_true_->GetXTitle(), sb.slice_vars_ );
+        vect_block->at(i).block_true_->GetXTitle(), sb.slice_vars_ );
       int yvar_idx = find_slice_var_index(
-        vect_block.at(i).block_true_->GetYTitle(), sb.slice_vars_ );
+        vect_block->at(i).block_true_->GetYTitle(), sb.slice_vars_ );
 
-      for( int j = 0; j < vect_block.at(i).block_true_->GetNBinsX(); j++ ){
-        double xlow = vect_block.at(i).block_true_->GetBinXLow(j);
-        double xhigh = vect_block.at(i).block_true_->GetBinXHigh(j);
-        auto& slice = add_slice( sb, vect_block.at(i).block_true_->GetVector(j),
+      for( int j = 0; j < vect_block->at(i).block_true_->GetNBinsX(); j++ ){
+        double xlow = vect_block->at(i).block_true_->GetBinXLow(j);
+        double xhigh = vect_block->at(i).block_true_->GetBinXHigh(j);
+        auto& slice = add_slice( sb, vect_block->at(i).block_true_->GetVector(j),
             xvar_idx, yvar_idx, xlow, xhigh );
-        for( int k = 0; k < vect_block.at(i).block_true_->GetNBinsY(j); k++ ){
+        for( int k = 0; k < vect_block->at(i).block_true_->GetNBinsY(j); k++ ){
           slice.bin_map_[ k + 1 ].insert( true_bins.size() );
-          true_bins.emplace_back(vect_block.at(i).block_true_->GetBinDef(j, k),
-            TrueBinType(vect_block.at(i).block_true_->GetBinType()), i );
-          reco_bins.emplace_back(vect_block.at(i).block_reco_->GetBinDef(j, k),
-            RecoBinType(vect_block.at(i).block_reco_->GetBinType()), i );
+          true_bins.emplace_back(vect_block->at(i).block_true_->GetBinDef(j, k),
+            TrueBinType(vect_block->at(i).block_true_->GetBinType()), i );
+          reco_bins.emplace_back(vect_block->at(i).block_reco_->GetBinDef(j, k),
+            RecoBinType(vect_block->at(i).block_reco_->GetBinType()), i );
         }
       }
     }
@@ -185,7 +179,6 @@ void MakeConfig::Print(){
       "" );
   int bin_number_var_idx = find_slice_var_index( "bin number", sb.slice_vars_ );
 
-
   // Create a slice showing all results together as a function of bin number
   auto& bin_num_slice = add_slice( sb, reco_bins.size(), 0, reco_bins.size(),
       bin_number_var_idx );
@@ -194,18 +187,18 @@ void MakeConfig::Print(){
     bin_num_slice.bin_map_[ ab + 1 ].insert( ab );
   }
 
-  for(const auto& bkg_idx : background_index){
-    std::string bdef = CATEGORY + Form( " == %d", bkg_idx );
-    true_bins.emplace_back( bdef, kBackgroundTrueBin, DUMMY_BLOCK_INDEX );
-  }
+  // Add a single true bin to collect background events by inverting the
+  // signal definition for the input selection
+  std::string bkgd_bdef = "!" + SELECTION + "_MC_Signal";
+  true_bins.emplace_back( bkgd_bdef, kBackgroundTrueBin, DUMMY_BLOCK_INDEX );
 
-  cout << DIRECTORY << '\n';
-  cout << TREE << '\n';
-  cout << SELECTION << '\n';
-  cout << true_bins.size() << '\n';
-  for ( const auto& tb : true_bins ) cout  << tb << '\n';
-  cout << reco_bins.size() << '\n';
-  for ( const auto& tb : reco_bins ) cout  << tb << '\n';
+  std::cout << DIRECTORY << '\n';
+  std::cout << TREE << '\n';
+  std::cout << SELECTION << '\n';
+  std::cout << true_bins.size() << '\n';
+  for ( const auto& tb : true_bins ) std::cout  << tb << '\n';
+  std::cout << reco_bins.size() << '\n';
+  for ( const auto& tb : reco_bins ) std::cout  << tb << '\n';
 
   std::string bin_config_output = std::getenv( "XSEC_ANALYZER_DIR" )
     + std::string( "/configs/" ) + BIN_CONFIG + "bin_config.txt";
@@ -224,7 +217,7 @@ void MakeConfig::Print(){
   std::string slice_config_output = std::getenv( "XSEC_ANALYZER_DIR" )
     + std::string( "/configs/" ) + BIN_CONFIG + "slice_config.txt";
 
-  cout << sb << '\n';
+  std::cout << sb << '\n';
   std::ofstream slice_out_file( slice_config_output );
   slice_out_file << sb;
   slice_out_file.close();
@@ -986,55 +979,24 @@ void Block2D::SetTexTitle( const std::string& textitle ) {
 
 void MakeConfig::BinScheme() {
 
+  // Instantiate the request binning scheme using the factory
+  BinSchemeFactory bsf;
+  BinSchemeBase* bs = bsf.CreateBinScheme( bin_scheme_name_ );
+  bin_scheme_.reset( bs );
+
   // The name of a TDirectoryFile which will store all of the histograms within
   // the output ROOT file
-  DIRECTORY = "tutorial_muon_1D_bin";
+  DIRECTORY = bin_scheme_->out_tdir_name_;
 
   // Prefix of output bin configure file and slice configure file
-  BIN_CONFIG = "tutorial_";
+  BIN_CONFIG = bin_scheme_->out_config_prefix_;
 
-  //
-  SELECTION = "CC1muXp0pi";
+  SELECTION = bin_scheme_->selection_name_;
+
+  TREE = bin_scheme_->ntuple_ttree_name_;
 
   // Runs used to plot smearing matrix
-  RUNS = { 1 };
-  vect_block.clear();
+  RUNS = bin_scheme_->runs_to_use_;
 
-  std::string branchexpr, title, textitle, selection;
-
-  // First block: cos_theta_mu in 1D
-  std::vector< double > cos_theta_mu_1D_edges = { -1., -0.85, -0.775, -0.7,
-    -0.625, -0.55, -0.475, -0.4, -0.325, -0.25, -0.175, -0.1, -0.025, 0.05,
-    0.125, 0.2, 0.275, 0.35, 0.425, 0.5, 0.575, 0.65, 0.725, 0.8, 0.85,
-    0.875, 0.9, 0.925, 0.950, 0.975, 1. };
-
-  Block1D* b1t = new Block1D( "CC1muXp0pi_true_muon_costh",
-    "muon cos#theta", "\\cos\\theta_{\\mu}", cos_theta_mu_1D_edges,
-    "CC1muXp0pi_MC_Signal", kSignalTrueBin );
-
-  Block1D* b1r = new Block1D( "CC1muXp0pi_reco_muon_costh",
-    "muon cos#theta", "\\cos\\theta_{\\mu}", cos_theta_mu_1D_edges,
-    "CC1muXp0pi_Selected", kOrdinaryRecoBin );
-
-  vect_block.emplace_back( b1t, b1r );
-
-  // Second block: p_mu in 1D
-  std::vector< double > pmu_1D_edges = { 0.1, 0.175, 0.2, 0.225, 0.25,
-    0.275, 0.3, 0.325, 0.35, 0.375, 0.4, 0.425, 0.45, 0.475, 0.5, 0.55, 0.6,
-    0.65, 0.7, 0.75, 0.8, 0.85, 0.9, 0.95, 1.0, 1.1, 1.2 };
-
-  Block1D* b2t = new Block1D( "CC1muXp0pi_true_muon_p",
-    "p_{#mu}; (GeV)", "p_{\\mu}; (GeV)", pmu_1D_edges,
-    "CC1muXp0pi_MC_Signal", kSignalTrueBin );
-
-  Block1D* b2r = new Block1D( "CC1muXp0pi_reco_muon_p",
-    "p_{#mu}; (GeV)", "p_{\\mu}; (GeV)", pmu_1D_edges,
-    "CC1muXp0pi_Selected", kOrdinaryRecoBin );
-
-  vect_block.emplace_back( b2t, b2r );
-
-  // CATEGORY is the branchexpr
-  // background_index is vector of background categories.
-  CATEGORY = "CC1muXp0pi_EventCategory";
-  background_index = {17, 18, 19, 20, 21, 22};
+  vect_block = &bin_scheme_->vect_block;
 }
